centered bg image on login page for small devices; refactored Layout component; hide sidebar by default on small devices
