+++ src/App.js
@@ -49,31 +49,15 @@ const AgentDeposit = Agent(Deposit2);
 Page404 = All(Page404);
 
 class App extends Component {
-	state = {
-		collapsed: false,
-	};
-
 	componentDidMount() {
 		// getContractsAddress();
 		// initBalanceProvider();
 		syncLoginState();
 	}
 
-	toggleSidebar = () => {
-		this.setState({
-			collapsed: !this.state.collapsed,
-		});
-	};
-
 	render() {
-		const { collapsed } = this.state;
-
 		return (
-			<Layout
-				isSideBarCollapsed={collapsed}
-				toggleSidebar={this.toggleSidebar}
-				simplified={["/login"]}
-			>
+			<Layout simplified={["/login"]}>
 				<Switch>
 					<Route path="/" exact component={Dashboard} />
 					<Route path="/login" exact component={Login} />
